+++ plaso/parsers/__init__.py
@@ -2,7 +2,6 @@
 """This file imports Python modules that register parsers."""
 
 from plaso.parsers import android_app_usage
-from plaso.parsers import android_logcat
 from plaso.parsers import asl
 from plaso.parsers import bash_history
 from plaso.parsers import bencode_parser
